+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 * Add CircleCI
 * Upgraded example Go implementation to show the use of custom environment variables
 * Build a Docker image on each merge to the develop branch
+* Accept a new Docker parameter, `SERVICES_RELATIVE_DIRPATH`, for the location (relative to the suite execution volume root) where file IO for the services created during test execution